Move DownloadGrid into Controls
