+++ src/NexusMods.App.UI/Controls/DownloadGrid/Columns/DownloadGameName/DownloadGameNameView.axaml
@@ -1,0 +1,22 @@
+﻿<reactiveUi:ReactiveUserControl
+    d:DesignHeight="21"
+    d:DesignWidth="120"
+    Width="140"
+    mc:Ignorable="d"
+    x:Class="NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadGameName.DownloadGameNameView"
+    x:TypeArguments="modName:IModNameViewModel"
+    xmlns="https://github.com/avaloniaui"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:reactiveUi="http://reactiveui.net"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:modName="clr-namespace:NexusMods.App.UI.RightContent.LoadoutGrid.Columns.ModName"
+    xmlns:downloadGameName="clr-namespace:NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadGameName">
+    <Design.DataContext>
+        <downloadGameName:DownloadGameNameDesignViewModel />
+    </Design.DataContext>
+    <TextBlock
+        VerticalAlignment="Center"
+        x:Name="GameTextBox" />
+</reactiveUi:ReactiveUserControl>
+
