+++ src/NexusMods.App.UI/Controls/DownloadGrid/Columns/DownloadGameName/DownloadGameNameViewModel.cs
@@ -1,0 +1,3 @@
+namespace NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadGameName;
+
+public class DownloadGameNameViewModel : DownloadGameNameDesignViewModel { }
