+++ src/NexusMods.App.UI/Controls/DownloadGrid/Columns/DownloadSize/DownloadSizeViewModel.cs
@@ -1,0 +1,4 @@
+namespace NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadSize;
+
+public class DownloadSizeViewModel : DownloadSizeDesignViewModel { }
+
