+++ src/NexusMods.App.UI/Controls/DownloadGrid/Columns/DownloadStatus/DownloadStatusView.axaml
@@ -1,0 +1,36 @@
+﻿<reactiveUi:ReactiveUserControl
+    d:DesignHeight="44"
+    d:DesignWidth="400"
+    mc:Ignorable="d"
+    x:Class="NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadStatus.DownloadStatusView"
+    x:TypeArguments="modName:IModNameViewModel"
+    xmlns="https://github.com/avaloniaui"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:reactiveUi="http://reactiveui.net"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:modName="clr-namespace:NexusMods.App.UI.RightContent.LoadoutGrid.Columns.ModName"
+    xmlns:resources="clr-namespace:NexusMods.App.UI.Resources"
+    xmlns:downloadStatus="clr-namespace:NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadStatus">
+    <Design.DataContext>
+        <downloadStatus:DownloadStatusDesignViewModel />
+    </Design.DataContext>
+
+    <StackPanel Orientation="Horizontal">
+
+        <ProgressBar Classes="DownloadBar"
+                     x:Name="DownloadProgressBar"
+                     Minimum="0" Maximum="1" ShowProgressText="True"
+                     ProgressTextFormat="{x:Static resources:Language.ProgressBar_ProgressTextFormat__Total_1_0}"
+                     Width="137" />
+
+        <Button
+            Height="24"
+            Width="24"
+            Padding="0"
+            Margin="18,0,18,0"
+            Classes="BareIcon PlayCircle"
+            x:Name="PlayPauseButton"/>
+
+    </StackPanel>
+</reactiveUi:ReactiveUserControl>
