+++ src/NexusMods.App.UI/Controls/DownloadGrid/Columns/DownloadStatus/DownloadStatusViewModel.cs
@@ -1,0 +1,3 @@
+namespace NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadStatus;
+
+public class DownloadStatusViewModel : DownloadStatusDesignViewModel { }
