+++ src/NexusMods.App.UI/Controls/DownloadGrid/Columns/DownloadVersion/DownloadVersionViewModel.cs
@@ -1,0 +1,3 @@
+namespace NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadVersion;
+
+public class DownloadVersionViewModel : DownloadVersionDesignViewModel { }
