+++ src/NexusMods.App.UI/Controls/DownloadGrid/DownloadColumn.cs
@@ -1,0 +1,13 @@
+namespace NexusMods.App.UI.Controls.DownloadGrid;
+
+/// <summary>
+/// Identifies the download grid columns.
+/// </summary>
+public enum DownloadColumn
+{
+    DownloadName,
+    DownloadVersion,
+    DownloadGameName,
+    DownloadSize,
+    DownloadStatus
+}
