--- src/NexusMods.App.UI/RightContent/DownloadGrid/Columns/DownloadGameName/DownloadGameNameViewModel.cs
@@ -1,3 +1,0 @@
-namespace NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadGameName;
-
-public class DownloadGameNameViewModel : DownloadGameNameDesignViewModel { }
