--- src/NexusMods.App.UI/RightContent/DownloadGrid/Columns/DownloadName/DownloadNameView.axaml
@@ -1,21 +1,0 @@
-﻿<reactiveUi:ReactiveUserControl
-    d:DesignHeight="44"
-    d:DesignWidth="400"
-    mc:Ignorable="d"
-    x:Class="NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadName.DownloadNameView"
-    x:TypeArguments="modName:IModNameViewModel"
-    xmlns="https://github.com/avaloniaui"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:reactiveUi="http://reactiveui.net"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:modName="clr-namespace:NexusMods.App.UI.RightContent.LoadoutGrid.Columns.ModName"
-    xmlns:downloadName1="clr-namespace:NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadName">
-    <Design.DataContext>
-        <downloadName1:DownloadNameDesignViewModel />
-    </Design.DataContext>
-    <TextBlock
-        VerticalAlignment="Center"
-        x:Name="NameTextBox" />
-</reactiveUi:ReactiveUserControl>
-
