--- src/NexusMods.App.UI/RightContent/DownloadGrid/Columns/DownloadSize/DownloadSizeViewModel.cs
@@ -1,4 +1,0 @@
-namespace NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadSize;
-
-public class DownloadSizeViewModel : DownloadSizeDesignViewModel { }
-
