--- src/NexusMods.App.UI/RightContent/DownloadGrid/Columns/DownloadSize/IDownloadSizeViewModel.cs
@@ -1,12 +1,0 @@
-﻿using NexusMods.App.UI.Controls.DataGrid;
-using NexusMods.App.UI.RightContent.Downloads.ViewModels;
-
-namespace NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadSize;
-
-/// <summary>
-/// Displays the name of a mod.
-/// </summary>
-public interface IDownloadSizeViewModel : IColumnViewModel<IDownloadTaskViewModel>
-{
-    public string Size { get; }
-}
