--- src/NexusMods.App.UI/RightContent/DownloadGrid/Columns/DownloadStatus/DownloadStatusViewModel.cs
@@ -1,3 +1,0 @@
-namespace NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadStatus;
-
-public class DownloadStatusViewModel : DownloadStatusDesignViewModel { }
