--- src/NexusMods.App.UI/RightContent/DownloadGrid/Columns/DownloadVersion/DownloadVersionView.axaml
@@ -1,22 +1,0 @@
-﻿<reactiveUi:ReactiveUserControl
-    d:DesignWidth="120"
-    d:DesignHeight="21"
-    Width="80"
-    mc:Ignorable="d"
-    x:Class="NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadVersion.DownloadVersionView"
-    x:TypeArguments="modName:IModNameViewModel"
-    xmlns="https://github.com/avaloniaui"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:reactiveUi="http://reactiveui.net"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:modName="clr-namespace:NexusMods.App.UI.RightContent.LoadoutGrid.Columns.ModName"
-    xmlns:downloadVersion1="clr-namespace:NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadVersion">
-    <Design.DataContext>
-        <downloadVersion1:DownloadVersionDesignViewModel />
-    </Design.DataContext>
-    <TextBlock
-        VerticalAlignment="Center"
-        x:Name="VersionTextBox" />
-</reactiveUi:ReactiveUserControl>
-
