--- src/NexusMods.App.UI/RightContent/DownloadGrid/Columns/DownloadVersion/DownloadVersionViewModel.cs
@@ -1,3 +1,0 @@
-namespace NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadVersion;
-
-public class DownloadVersionViewModel : DownloadVersionDesignViewModel { }
