--- src/NexusMods.App.UI/RightContent/DownloadGrid/DownloadColumn.cs
@@ -1,13 +1,0 @@
-namespace NexusMods.App.UI.RightContent.DownloadGrid;
-
-/// <summary>
-/// Identifies the download grid columns.
-/// </summary>
-public enum DownloadColumn
-{
-    DownloadName,
-    DownloadVersion,
-    DownloadGameName,
-    DownloadSize,
-    DownloadStatus
-}
