+++ src/NexusMods.App.UI/RightContent/Downloads/InProgressViewModel.cs
@@ -8,13 +8,13 @@ using DynamicData;
 using DynamicData.Binding;
 using JetBrains.Annotations;
 using NexusMods.App.UI.Controls.DataGrid;
+using NexusMods.App.UI.Controls.DownloadGrid;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadGameName;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadName;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadSize;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadStatus;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadVersion;
 using NexusMods.App.UI.Overlays;
-using NexusMods.App.UI.RightContent.DownloadGrid;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadGameName;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadName;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadSize;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadStatus;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadVersion;
 using NexusMods.App.UI.RightContent.Downloads.ViewModels;
 using NexusMods.Networking.Downloaders.Interfaces;
 using ReactiveUI;
