+++ src/NexusMods.App.UI/Services.cs
@@ -5,6 +5,11 @@ using NexusMods.Abstractions.Serialization.Json;
 using NexusMods.App.UI.Controls;
 using NexusMods.App.UI.Controls.DataGrid;
 using NexusMods.App.UI.Controls.DevelopmentBuildBanner;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadGameName;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadName;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadSize;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadStatus;
+using NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadVersion;
 using NexusMods.App.UI.Controls.GameWidget;
 using NexusMods.App.UI.Controls.Spine;
 using NexusMods.App.UI.Controls.Spine.Buttons.Download;
@@ -23,11 +28,6 @@ using NexusMods.App.UI.Overlays.Login;
 using NexusMods.App.UI.Overlays.MetricsOptIn;
 using NexusMods.App.UI.Overlays.Updater;
 using NexusMods.App.UI.RightContent;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadGameName;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadName;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadSize;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadStatus;
-using NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadVersion;
 using NexusMods.App.UI.RightContent.Downloads;
 using NexusMods.App.UI.RightContent.LoadoutGrid;
 using NexusMods.App.UI.RightContent.LoadoutGrid.Columns.ModCategory;
@@ -40,11 +40,11 @@ using NexusMods.App.UI.Windows;
 using NexusMods.App.UI.WorkspaceSystem;
 using NexusMods.Paths;
 using ReactiveUI;
-using DownloadGameNameView = NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadGameName.DownloadGameNameView;
-using DownloadNameView = NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadName.DownloadNameView;
-using DownloadSizeView = NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadSize.DownloadSizeView;
-using DownloadStatusView = NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadStatus.DownloadStatusView;
-using DownloadVersionView = NexusMods.App.UI.RightContent.DownloadGrid.Columns.DownloadVersion.DownloadVersionView;
+using DownloadGameNameView = NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadGameName.DownloadGameNameView;
+using DownloadNameView = NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadName.DownloadNameView;
+using DownloadSizeView = NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadSize.DownloadSizeView;
+using DownloadStatusView = NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadStatus.DownloadStatusView;
+using DownloadVersionView = NexusMods.App.UI.Controls.DownloadGrid.Columns.DownloadVersion.DownloadVersionView;
 using ImageButton = NexusMods.App.UI.Controls.Spine.Buttons.Image.ImageButton;
 using ModCategoryView = NexusMods.App.UI.RightContent.LoadoutGrid.Columns.ModCategory.ModCategoryView;
 using ModEnabledView = NexusMods.App.UI.RightContent.LoadoutGrid.Columns.ModEnabled.ModEnabledView;
